Correct docs on PAS cert rotation for IDPs
The updated structure is much more readable and calls attention to need to update SSO service.

However, the previous instructions were taken from an SSO Service KB and not translated correctly to apply to PAS, but enough to no longer work for the SSO service. As a result, the steps wouldn't work for both PAS and the SSO service, nor does it correctly represent the conditions under which the steps must be performed.

Modified the instructions accordingly to reflect both things must be changed at the same time as a result of triggering PAS certificate rotation. Also fixed the instructions to work in the context of UAA, as well as broken links to SSO integration guides.

And proactively going to mention that the steps should be included here so the steps for PAS cert rotation are self-contained and **customers know what steps to take as part of the PAS process to rotate certificates**.

And proactively going to mention `integration (a.k.a. trust or partnernship or Federation)` is not a mistake, it is known by many different names as each IDP tends to call it something else.

This change should be backported to all supported versions of PAS.