Update docs to match new behavior in 2.9
1. Since BOSH DNS is now rotatable via CredHub Maestro, I've updated the
docs to change how we label that specific certificate.

2. Also, since OpsManager now rotates CredHub certificates, it is
essential to inform platform operators that service instances must be
upgraded as well as recreated in order to correctly push certificates to
each service instance. I also clarified the effect of each errand.

3. Lastly, there are is a list of non-rotatable certificates controlled by
OpsManager. I wanted to list them to make it easier for platform
operators to check.

[#171161837](https://www.pivotaltracker.com/story/show/171161837)